requirements.txt
openai 
python-dotenv 
SpeechRecognition 
pyttsx3 
gtts 
wheel
PyAudio
numpy